+++ datafusion/physical-expr-common/Cargo.toml
@@ -36,7 +36,13 @@ name = "datafusion_physical_expr_common"
 path = "src/lib.rs"
 
 [dependencies]
+ahash = { version = "0.8", default-features = false, features = [
+    "runtime-rng",
+] }
 arrow = { workspace = true }
+arrow-schema = { workspace = true }
 datafusion-common = { workspace = true, default-features = true }
 datafusion-expr = { workspace = true }
+half = { workspace = true }
+hashbrown = { version = "0.14", features = ["raw"] }
 paste = "^1.0"
